delete more unused classes
