+++ core/src/main/resources/reference.conf
@@ -128,13 +128,6 @@ akka.http.server.idle-timeout = 210 s
 akka.http.server.request-timeout=180 s
 akka.http.server.server-header = ""
 
-dataRepoEntityProvider {
-  # 10 expressions * 100000 base table rows
-  maxInputsPerSubmission = 1000000
-  maxBigQueryResponseSize = 500MB
-  bigQueryMaximumBytesBilled = 100GB
-}
-
 leonardo {
     wdsType = "WDS" # "CROMWELL" for joint WDS+CBAS; "WDS" for standalone/decoupled WDS
     server = "add-leo-hostname-here"
